feat: use zact for server actions
